Split gather functions, adding utils
This adds one new platform-specific module for FreeBSD network programs, and
two new utils that deal with python sockets (well supported on all platforms)
and the interfaces data-structure.
