+++ salt/modules/freebsd_network.py
@@ -1,0 +1,188 @@
+'''
+Module for gathering and managing network information
+'''
+
+from salt.utils.interfaces import *
+
+__outputter__ = {
+    'dig':     'txt',
+    'ping':    'txt',
+    'netstat': 'txt',
+}
+
+def __virtual__():
+    '''
+    Only works on *BSD systems
+    '''
+    if __grains__['os'] in ('FreeBSD', 'NetBSD', 'OpenBSD'):
+        return 'network'
+    return False
+
+def _interfaces_ifconfig(out):
+    '''
+    Uses ifconfig to return a dictionary of interfaces with various information
+    about each (up/down state, ip address, netmask, and hwaddr)
+    '''
+    import re
+    ret = dict()
+
+    piface = re.compile('^(\S+):?')
+    popts = re.compile('\s*(?:options=)([0-9a-fA-F])<([^>]*)>')
+    pmac = re.compile('.*?(?:HWaddr|ether) ([0-9a-fA-F:]+)')
+    pip = re.compile('.*?(?:inet addr:|inet )(.*?)\s')
+    pip6 = re.compile('.*?(?:inet6 addr: (.*?)/|inet6 )([0-9a-fA-F:]+)')
+    pmask = re.compile('.*?(?:Mask:|netmask )(?:(0x[0-9a-fA-F]{8})|([\d\.]+))')
+    pmask6 = re.compile('.*?(?:inet6 addr: [0-9a-fA-F:]+/(\d+)|prefixlen (\d+)).*')
+    pupdown = re.compile('UP')
+    pbcast = re.compile('.*?(?:Bcast:|broadcast )([\d\.]+)')
+
+    groups = re.compile('\r?\n(?=\S)').split(out)
+    for group in groups:
+        data = dict()
+        iface = ''
+        updown = False
+        for line in group.split('\n'):
+            miface = piface.match(line)
+            mopts = popts.match(line)
+            mmac = pmac.match(line)
+            mip = pip.match(line)
+            mip6 = pip6.match(line)
+            mupdown = pupdown.search(line)
+            if miface:
+                iface = miface.group(1)
+            if mopts:
+                data['flags'] = mopts.group(1)
+                data['caps']  = mopts.group(2)
+            if mmac:
+                data['hwaddr'] = mmac.group(1)
+            if mip:
+                if 'inet' not in data:
+                    data['inet'] = list()
+                addr_obj = dict()
+                addr_obj['address'] = mip.group(1)
+                mmask = pmask.match(line)
+                if mmask:
+                    if mmask.group(1):
+                        mmask = int(mmask.group(1), 16)
+                    else:
+                        mmask = mmask.group(2)
+                    addr_obj['netmask'] = mmask
+                mbcast = pbcast.match(line)
+                if mbcast:
+                    addr_obj['broadcast'] = mbcast.group(1)
+                data['inet'].append(addr_obj)
+            if mupdown:
+                updown = True
+            if mip6:
+                if 'inet6' not in data:
+                    data['inet6'] = list()
+                addr_obj = dict()
+                addr_obj['address'] = mip6.group(1) or mip6.group(2)
+                mmask6 = pmask6.match(line)
+                if mmask6:
+                    addr_obj['prefixlen'] = mmask6.group(1) or mmask6.group(2)
+                data['inet6'].append(addr_obj)
+        data['up'] = updown
+        ret[iface] = data
+        del data
+    return ret
+
+
+def interfaces():
+    ifaces = dict()
+    cmd = __salt__['cmd.run']('ifconfig -a')
+    ifaces = _interfaces_ifconfig(cmd)
+    return ifaces
+
+
+def ping(host):
+    '''
+    Performs a ping to a host
+
+    CLI Example::
+
+        salt '*' network.ping archlinux.org
+    '''
+    cmd = 'ping -c 4 %s' % _sanitize_host(host)
+    return __salt__['cmd.run'](cmd)
+
+
+def netstat():
+    '''
+    Return information on open ports and states
+
+    CLI Example::
+
+        salt '*' network.netstat
+    '''
+    ret = []
+    cmd = 'netstat -tulpnea'
+    out = __salt__['cmd.run'](cmd).split('\n')
+    for line in out:
+        comps = line.split()
+        if line.startswith('tcp'):
+            ret.append({
+                'inode': comps[7],
+                'local-address': comps[3],
+                'program': comps[8],
+                'proto': comps[0],
+                'recv-q': comps[1],
+                'remote-address': comps[4],
+                'send-q': comps[2],
+                'state': comps[5],
+                'user': comps[6]})
+        if line.startswith('udp'):
+            ret.append({
+                'inode': comps[6],
+                'local-address': comps[3],
+                'program': comps[7],
+                'proto': comps[0],
+                'recv-q': comps[1],
+                'remote-address': comps[4],
+                'send-q': comps[2],
+                'user': comps[5]})
+    return ret
+
+
+def traceroute(host):
+    '''
+    Performs a traceroute to a 3rd party host
+
+    CLI Example::
+
+        salt '*' network.traceroute archlinux.org
+    '''
+    ret = []
+    cmd = 'traceroute %s' % _sanitize_host(host)
+    out = __salt__['cmd.run'](cmd)
+
+    for line in out:
+        if not ' ' in line:
+            continue
+        if line.startswith('traceroute'):
+            continue
+        comps = line.split()
+        result = {
+            'count': comps[0],
+            'hostname': comps[1],
+            'ip': comps[2],
+            'ms1': comps[4],
+            'ms2': comps[6],
+            'ms3': comps[8],
+            'ping1': comps[3],
+            'ping2': comps[5],
+            'ping3': comps[7]}
+        ret.append(result)
+    return ret
+
+
+def dig(host):
+    '''
+    Performs a DNS lookup with dig
+
+    CLI Example::
+
+        salt '*' network.dig archlinux.org
+    '''
+    cmd = 'dig %s' % _sanitize_host(host)
+    return __salt__['cmd.run'](cmd)
